requirements.txt
aliyun-log-python-sdk==0.9.0
Brotli
certifi
charset-normalizer
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
dateparser==1.2.0
elastic-transport==8.12.0
elasticsearch==8.12.1
flatbuffers==23.5.26
fonttools==4.51.0
humanfriendly==10.0
idna
imageio==2.34.1
imgaug==0.4.0
jmespath==1.0.1
kiwisolver==1.4.5
lazy_loader==0.4
markdown-it-py==3.0.0
matplotlib==3.8.4
mdurl==0.1.2
MNN==2.8.3
mpmath==1.3.0
networkx==3.3
numpy
onnx==1.15.0
onnx-tool==0.9.0
onnxruntime==1.17.1
onnxsim==0.4.36
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.1
pillow
protobuf==3.20.3
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
PySocks
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2023.12.25
requests 
rich==13.7.1
scikit-image==0.23.2
scipy==1.13.0
shapely==2.0.4
six==1.16.0
sympy==1.12
tabulate==0.9.0
tifffile==2024.5.10
torch==1.12.1
torchaudio==0.12.1
torchsummary==1.5.1
torchvision==0.13.1
tqdm==4.66.2
typing_extensions
tzdata==2024.1
tzlocal==5.2
urllib3
